backend/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
firebase-admin==6.3.0
Flask==3.0.0
Flask-Cors==4.0.0
flower==2.0.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.3
msgpack==1.0.7
packaging==23.2
pillow==10.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyminizip==0.2.6
PyMuPDF==1.23.15
PyMuPDFb==1.23.9
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.39.2
six==1.16.0
tornado==6.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
